Don't use webhooks for patching websites in load tests
Webhooks add a tremendous amount of latency to API calls, especially during load tests when the kyverno admission controller is overloaded.
High latency of API calls extremely decreases the operator's reconciliation rate which makes it difficult to achieve the desired load.
We are not load testing kyverno or the API server's webhook machinery, so get that off the critical path.