Stop using state label for anything else than observability
We don't want to rely on the state label being maintained corretly in the sharder webhook or controller.
Determining the state directly based on the local clock is favorable as it makes it easier to trigger in predicates when the controller will actually act on changed lease data.
This increases the rate of ring calculations again, but is more consistent and correct.
This was a premature optimization. We can still optimize ring calculations again when it becomes real problem.